Modular refactor
